pyproject.toml
[tool.poetry]
name = "robotframework-openpdf"
version = "0.1.0"
description = ""
authors = ["Bruno Moraes <moraes.7bruno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.1"
tika = "^1.24"
robotframework = "^3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
